migration fix one more time

+++ database/migrations/2020_12_28_103405_create_category_user_table.php
@@ -14,7 +14,7 @@ class CreateCategoryUserTable extends Migration
     public function up()
     {
         Schema::create('category_user', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedInteger('user_id')->index();
             $table->unsignedInteger('category_id')->index();
             $table->timestamps();
